requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
tqdm==4.65.0
emoji==2.10.1 